fix tests, remove Route and marsTest (now irrelevant/depreciated tests)
